+++ GaggiaPIDController/libraries/Factories.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "BaseTypes.h"
-#include "Configuration.h"
 #include "Sensors.h"
 #include "TemperatureSensor.h"
+#include "coffee_machine/BaseTypes.h"
+#include "coffee_machine/Configuration.h"
 
 class SensorFactory
 {