+++ GaggiaPIDController/libraries/Sensors.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "BaseTypes.h"
-#include "Configuration.h"
+#include "coffee_machine/BaseTypes.h"
+#include "coffee_machine/Configuration.h"
 
 #include "max6675.h"
 #include <TSIC.h>