+++ scripts/test-project.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+
+set -euo pipefail
+
+# Build the automated test suite
+mkdir -p build
+cd build
+cmake ..
+make
+
+# Run the test
+cd test
+export GTEST_COLOR=1
+ctest --verbose